Move common imports to top-level; fix model saving bugs.
