+++ examples/data/synthesis/synthesize_passbys.py
@@ -7,8 +7,7 @@ import sys
 import matplotlib.pyplot as plt
 import numpy as np
 
-from riid.data.synthetic import get_dummy_seeds
-from riid.data.synthetic.passby import PassbySynthesizer
+from riid import PassbySynthesizer, get_dummy_seeds
 
 if len(sys.argv) == 2:
     import matplotlib
