+++ examples/data/synthesis/synthesize_seeds.py
@@ -4,7 +4,7 @@
 """This example demonstrates how to generate synthetic seeds from GADRAS."""
 import yaml
 
-from riid.data.synthetic.seed import SeedSynthesizer
+from riid import SeedSynthesizer
 
 seed_synth_config = """
 ---
