+++ examples/modeling/label_proportion_estimation.py
@@ -4,10 +4,8 @@ import os
 
 from sklearn.metrics import mean_absolute_error
 
-from riid.data.synthetic import get_dummy_seeds
-from riid.data.synthetic.seed import SeedMixer
-from riid.data.synthetic.static import StaticSynthesizer
-from riid.models.neural_nets import LabelProportionEstimator
+from riid import SeedMixer, StaticSynthesizer, get_dummy_seeds
+from riid.models import LabelProportionEstimator
 
 # Generate some mixture training data.
 fg_seeds_ss, bg_seeds_ss = get_dummy_seeds().split_fg_and_bg()
