+++ examples/run_examples.py
@@ -6,6 +6,7 @@ import os
 import subprocess
 import sys
 from pathlib import Path
+
 import pandas as pd
 from tabulate import tabulate
