+++ tests/data_tests.py
@@ -6,10 +6,10 @@ import os
 import tempfile
 import unittest
 
-from riid import SAMPLESET_HDF_FILE_EXTENSION
+from riid import (SAMPLESET_HDF_FILE_EXTENSION, SampleSet, get_dummy_seeds,
+                  read_hdf)
 from riid.data.labeling import label_to_index_element
-from riid.data.sampleset import SampleSet, _write_hdf, read_hdf
-from riid.data.synthetic import get_dummy_seeds
+from riid.data.sampleset import _write_hdf
 
 
 class TestData(unittest.TestCase):
