+++ tests/staticsynth_tests.py
@@ -7,12 +7,11 @@ import unittest
 import numpy as np
 import pandas as pd
 
+from riid import SeedMixer, StaticSynthesizer, get_dummy_seeds
 from riid.data import InvalidSeedError, get_expected_spectra
-from riid.data.synthetic import (Synthesizer, get_dummy_seeds,
-                                 get_merged_sources_samplewise,
-                                 get_samples_per_seed)
-from riid.data.synthetic.seed import SeedMixer
-from riid.data.synthetic.static import StaticSynthesizer
+from riid.data.synthetic.base import (Synthesizer,
+                                      get_merged_sources_samplewise,
+                                      get_samples_per_seed)
 
 
 class TestStaticSynthesis(unittest.TestCase):
