+++ tests/visualize_tests.py
@@ -6,11 +6,9 @@ import unittest
 
 import numpy as np
 
-from riid.data.synthetic import get_dummy_seeds
-from riid.data.synthetic.seed import SeedMixer
-from riid.data.synthetic.static import StaticSynthesizer
+from riid import SeedMixer, StaticSynthesizer, get_dummy_seeds
 from riid.metrics import precision_recall_curve
-from riid.models.neural_nets import MLPClassifier
+from riid.models import MLPClassifier
 from riid.visualize import (plot_correlation_between_all_labels,
                             plot_count_rate_history,
                             plot_label_and_prediction_distributions,
